feat(Topico de aula completas): :memo: 04. Fazendo requisições para web
Ensinado sobre retornar mais de uma resposta em uma função, e de usar estas respostas, ensinado sobre o loop for, e requisição para web com pacote

Organizado as pastas e arquivos
