+++ alura/app/06.lendo-arquivos/arquivos-em-GO.go
