+++ alura/app/06.lendo-arquivos/detectando-erros.go
