+++ alura/app/06.lendo-arquivos/log.txt
