+++ alura/app/topicos_estudo/03.fluxo-switch.go
@@ -2,13 +2,14 @@ package main
 
 import (
 	"fmt"
+	"os"
 )
 
-func main () {
+func main() {
 	fmt.Println("1- Iniciar Monitoramento")
 	fmt.Println("2- Exibir Logs")
 	fmt.Println("0- Sair do Programa")
-	
+
 	var comando int
 	fmt.Scanf("%d", &comando)
 
@@ -25,14 +26,15 @@ func main () {
 	// }
 
 	switch comando {
-		case 1:
-			fmt.Println("Monitorando...")
-		case 2:
-			fmt.Println("Exibindo Logs...")
-		case 0:
-			fmt.Println("Saindo do Programa !")
-		default:
-			fmt.Println("Comando invalido")
+	case 1:
+		fmt.Println("Monitorando...")
+	case 2:
+		fmt.Println("Exibindo Logs...")
+	case 0:
+		fmt.Println("Saindo do Programa !")
+		os.Exit(0)
+	default:
+		fmt.Println("Comando invalido")
 	}
 
 	// No GO não precisa colocar o break
