+++ alura/parte_2/banco/02.referencia-ponteiro-metodo/metodo-sacar.go
